+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog
 - added `serialize`, `deserialize` methods to `weka.filters.Filter` to simplify loading/saving filter
 - added methods `plot_rocs` and `plot_prcs` to `weka.plot.classifiers` module to plot ROC/PRC curve on same dataset
   for multiple classifiers
+- method `create_instances_from_matrices` from module `weka.core.dataset` now allows string and bytes as well
 - ...